# Fix 20x slowdown of FP6 kernel due to device properties query