@gau-nernst 2024-10-17
The numbers look good already. But I wonder if we can cache this result? (e.g. use static variable or global variable - I know, frowned upon by C++ devs :rofl:). Will using PyTorch API be better? (i.e. `torch.cuda.get_device_capability()` - and does this function cache the result?)

(Again, the current state looks good already. Just something I wonder)