@tobiasvanderwerff 2024-10-17
@gau-nernst just FYI, FlashAttention uses `at::cuda::getCurrentDeviceProperties();` ([source](
https://github.com/Dao-AILab/flash-attention/blob/c1d146cbd5becd9e33634b1310c2d27a49c7e862/hopper/flash_api.cpp#L477-L478)), which I also considered. However, I couldn't find documentation on whether it has the same problem (querying all device properties at once, which is slow), or that it handles this in some more optimized way.